Админ панель V2.0

+++ src/main/java/com/example/demo/repository/ProductRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
+    // Здесь можно добавить дополнительные методы для поиска, если это необходимо
 }
